refactor(project-restructure): :art: Rename utilities
